+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+### 1.5.1
+
+- fix marshalling of options to SMILES
+
+### 1.5.0
+
+- fix [#387](https://github.com/PDConSec/vsc-print/issues/387)
+- fix SMILES and support config (see manual)
+
 ### 1.4.1
 
 - fix [#386](https://github.com/PDConSec/vsc-print/issues/386)